Initial work on safer termbox bindings with Rust/rustler
Tests are passing already, but I'd like to further investigate how
termbox crashes are handled now.

Still need to automate building the termbox library and configuring the
location of the built static lib (via a `build.rs`). Currently, it's
possible to manually copy `libtermbox.a` into e.g. `target/debug/deps/`.

See #2.

+++ c_src/termbox_bindings.c
@@ -136,7 +136,7 @@ static ERL_NIF_TERM extb_set_clear_attributes(ErlNifEnv *env, int argc,
   enif_get_uint(env, argv[0], &fg);
   enif_get_uint(env, argv[1], &bg);
 
-  tb_set_clear_attributes((uint32_t)fg, (uint32_t)bg);
+  tb_set_clear_attributes((uint16_t)fg, (uint16_t)bg);
   return extb_ok(env);
 }
